# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.Housing:
    properties:
      city:
        example: Lille
        type: string
      country:
        example: FR
        type: string
      has_electricity:
        example: true
        type: boolean
      has_gas:
        example: false
        type: boolean
      is_furnished:
        example: false
        type: boolean
      last_tenant_id:
        example: e185deb2-91d5-4ab7-87b3-daaffac00e3d
        type: string
      owner_id:
        example: e185deb2-91d5-4ab7-87b3-daaffac00e3d
        type: string
      rent_price:
        example: 60.95
        type: number
      rental_charges:
        example: 60.9
        type: number
      state:
        example: Haut-de-France
        type: string
      status_id:
        example: e185deb2-91d5-4ab7-87b3-daaffac00e3d
        type: string
      street:
        example: 78 Rue Solférino
        type: string
      surface_area:
        example: 15.5
        type: number
      type_id:
        example: e185deb2-91d5-4ab7-87b3-daaffac00e3d
        type: string
      zip:
        example: "59000"
        type: string
    type: object
  models.HousingBody:
    properties:
      city:
        example: Lille
        type: string
      country:
        example: FR
        type: string
      has_electricity:
        example: true
        type: boolean
      has_gas:
        example: false
        type: boolean
      is_furnished:
        example: false
        type: boolean
      last_tenant_id:
        example: e185deb2-91d5-4ab7-87b3-daaffac00e3d
        type: string
      owner_id:
        example: e185deb2-91d5-4ab7-87b3-daaffac00e3d
        type: string
      rent_price:
        example: 60.95
        type: number
      rental_charges:
        example: 60.9
        type: number
      state:
        example: Haut-de-France
        type: string
      status_id:
        example: e185deb2-91d5-4ab7-87b3-daaffac00e3d
        type: string
      street:
        example: 78 Rue Solférino
        type: string
      surface_area:
        example: 15.5
        type: number
      type_id:
        example: e185deb2-91d5-4ab7-87b3-daaffac00e3d
        type: string
      zip:
        example: "59000"
        type: string
    type: object
  models.HousingType:
    properties:
      name:
        example: garage
        type: string
    type: object
  models.Status:
    properties:
      name:
        example: sold
        type: string
    type: object
  models.UpdateStatusBody:
    properties:
      status_id:
        example: e185deb2-91d5-4ab7-87b3-daaffac00e3d
        type: string
    type: object
info:
  contact: {}
  license: {}
  title: Housing API
  version: 0.1.0
paths:
  /housing:
    get:
      description: Search all housing
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Get all housing
      tags:
      - Housing
    post:
      consumes:
      - application/json
      description: Create a new housing with given data
      parameters:
      - description: Housing data
        in: body
        name: housing
        required: true
        schema:
          $ref: '#/definitions/models.Housing'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "422":
          description: Unprocessable Entity
          schema:
            type: string
      summary: Insert a housing
      tags:
      - Housing
  /housing/{housing_id}:
    delete:
      description: Delete a given housing by ID
      parameters:
      - description: Housing ID
        in: path
        name: housing_id
        required: true
        type: string
      responses:
        "204": {}
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Delete a housing by ID
      tags:
      - Housing
    get:
      description: Search for a given housing with its ID
      parameters:
      - description: Housing ID
        in: path
        name: housing_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Get a housing by ID
      tags:
      - Housing
    put:
      consumes:
      - application/json
      description: Update a given housing by ID
      parameters:
      - description: Housing ID
        in: path
        name: housing_id
        required: true
        type: string
      - description: Housing data
        in: body
        name: housing
        required: true
        schema:
          $ref: '#/definitions/models.HousingBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "422":
          description: Unprocessable Entity
          schema:
            type: string
      summary: Update a housing by ID
      tags:
      - Housing
  /housing/{housing_id}/status:
    put:
      consumes:
      - application/json
      description: Update the status of a given housing
      parameters:
      - description: Housing ID
        in: path
        name: housing_id
        required: true
        type: string
      - description: Status ID
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/models.UpdateStatusBody'
      responses:
        "204": {}
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "422":
          description: Unprocessable Entity
          schema:
            type: string
      summary: Update status of housing
      tags:
      - Housing
  /housing/status:
    get:
      description: Search all housing statuses
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Get all housing statuses
      tags:
      - Housing Status
    post:
      consumes:
      - application/json
      description: Create a new housing status with given data
      parameters:
      - description: Housing status data
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/models.Status'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "422":
          description: Unprocessable Entity
          schema:
            type: string
      summary: Insert a housing status
      tags:
      - Housing Status
  /housing/type:
    get:
      description: Search all housing types
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Get all housing types
      tags:
      - Housing Types
    post:
      consumes:
      - application/json
      description: Create a new housing type with given data
      parameters:
      - description: Housing type data
        in: body
        name: type
        required: true
        schema:
          $ref: '#/definitions/models.HousingType'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "422":
          description: Unprocessable Entity
          schema:
            type: string
      summary: Insert a housing type
      tags:
      - Housing Types
  /housing/type/{housing_type_id}:
    get:
      description: Search all housing of a given type
      parameters:
      - description: Housing type ID
        in: path
        name: housing_type_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: Get all housing by type
      tags:
      - Housing Types
swagger: "2.0"